refactor(mixins): split QueueRuntimeConfigMixin from decorators
register_type, set_queue_rate_limit, and set_queue_concurrency are
runtime configuration knobs, not decorator surface. Moving them to a
dedicated QueueRuntimeConfigMixin pulls 64 LOC out of decorators.py
(597 -> 533) and gives the three queue-tuning APIs a coherent home
adjacent to the existing QueueSettingsMixin (dashboard key/value store).
