+++ py_src/taskito/mixins/decorators.py
@@ -18,7 +18,6 @@ from taskito.context import _clear_context, current_job
 from taskito.events import EventType
 from taskito.inject import Inject, _InjectAlias
 from taskito.interception.reconstruct import reconstruct_args
-from taskito.interception.strategy import Strategy as S
 from taskito.predicates.core import coerce_predicate
 from taskito.proxies import cleanup_proxies, reconstruct_proxies
 from taskito.task import TaskWrapper
@@ -531,66 +530,3 @@ class QueueDecoratorMixin:
         """
         self._hooks["on_failure"].append(fn)
         return fn
-
-    # -- Type registration --
-
-    def register_type(
-        self,
-        python_type: type,
-        strategy: str,
-        *,
-        resource: str | None = None,
-        message: str | None = None,
-        converter: Callable | None = None,
-        type_key: str | None = None,
-        proxy_handler: str | None = None,
-    ) -> None:
-        """Register a custom type with the interception system.
-
-        Args:
-            python_type: The type to register.
-            strategy: One of ``"pass"``, ``"convert"``, ``"redirect"``,
-                ``"reject"``, or ``"proxy"``.
-            resource: Resource name for ``"redirect"`` strategy.
-            message: Rejection reason for ``"reject"`` strategy.
-            converter: Converter callable for ``"convert"`` strategy.
-            type_key: Key for the converter reconstructor dispatch.
-            proxy_handler: Handler name for ``"proxy"`` strategy.
-        """
-        if self._interceptor is None:
-            raise RuntimeError(
-                "Interception is disabled; set interception='strict' or "
-                "'lenient' to use register_type()"
-            )
-        strat = S(strategy)
-        self._interceptor._registry.register(
-            python_type,
-            strat,
-            priority=15,
-            redirect_resource=resource,
-            reject_reason=message,
-            converter=converter,
-            type_key=type_key,
-            proxy_handler=proxy_handler,
-        )
-
-    # -- Queue-level config --
-
-    def set_queue_rate_limit(self, queue_name: str, rate_limit: str) -> None:
-        """Set a rate limit for an entire queue.
-
-        Args:
-            queue_name: Queue name (e.g. ``"default"``).
-            rate_limit: Rate limit string, e.g. ``"100/m"``, ``"10/s"``.
-        """
-        self._queue_configs.setdefault(queue_name, {})["rate_limit"] = rate_limit
-
-    def set_queue_concurrency(self, queue_name: str, max_concurrent: int) -> None:
-        """Set a maximum number of concurrent jobs for a queue.
-
-        Args:
-            queue_name: Queue name (e.g. ``"default"``).
-            max_concurrent: Maximum number of jobs running simultaneously
-                from this queue.
-        """
-        self._queue_configs.setdefault(queue_name, {})["max_concurrent"] = max_concurrent
